Первый коммит с реализацией финального проекта 12 спринта.

+++ src/test/java/ru/yandex/practicum/filmorate/validation/FilmValidationTest.java
@@ -29,7 +29,7 @@ class FilmValidationTest {
         film.setName("Film test");
         film.setDescription("Description test");
         film.setReleaseDate(LocalDate.of(1997, 12, 31));
-        film.setDuration(100L);
+        film.setDuration(100);
     }
 
     @DisplayName("Валидация поля name")
@@ -105,7 +105,7 @@ class FilmValidationTest {
             "1, OK",
             "100, OK"
     })
-    void durationValidation(long duration, String expectedErrorMessage) {
+    void durationValidation(Integer duration, String expectedErrorMessage) {
         film.setDuration(duration);
         List<String> violations = validator.validate(film).stream()
                 .map(ConstraintViolation::getMessage)
